Refactor: 리뷰 요청사항 반영 (#64)
* Fix: 카메라 촬영 액티비티 불필요한 코드 제거

* Refactor: DI 파일 이동

* Refactor: Lint 수행
