+++ app/src/main/java/com/kappzzang/jeongsan/di/RepositoryModule.kt
@@ -1,23 +1,13 @@
 package com.kappzzang.jeongsan.di
 
-import com.kappzzang.jeongsan.repository.ExpenseDetailRepository
-import com.kappzzang.jeongsan.repository.ExpenseListRepository
-import com.kappzzang.jeongsan.repository.ExpenseRepository
 import com.kappzzang.jeongsan.repository.GroupInfoRepository
 import com.kappzzang.jeongsan.repository.KakaoAuthenticationRepository
 import com.kappzzang.jeongsan.repository.MemberRepository
-import com.kappzzang.jeongsan.repository.ReceiptRepository
-import com.kappzzang.jeongsan.repository.TransferRepository
 import com.kappzzang.jeongsan.repository.UserInfoRepository
 import com.kappzzang.jeongsan.repositoryimpl.AuthenticationRepositoryImpl
-import com.kappzzang.jeongsan.repositoryimpl.ExpenseDetailRepositoryImpl
-import com.kappzzang.jeongsan.repositoryimpl.ExpenseListFakeRepositoryImpl
-import com.kappzzang.jeongsan.repositoryimpl.ExpenseRepositoryImpl
 import com.kappzzang.jeongsan.repositoryimpl.GroupInfoRepositoryImpl
 import com.kappzzang.jeongsan.repositoryimpl.KakaoAuthenticationRepositoryImpl
 import com.kappzzang.jeongsan.repositoryimpl.MemberRepositoryImpl
-import com.kappzzang.jeongsan.repositoryimpl.ReceiptRepositoryImpl
-import com.kappzzang.jeongsan.repositoryimpl.TransferRepositoryImpl
 import com.kappzzang.jeongsan.repositoryimpl.UserInfoRepositoryImpl
 import com.kappzzang.jeongsan.util.AuthenticationRepository
 import dagger.Binds
@@ -44,31 +34,7 @@ abstract class RepositoryModule {
 
     @Binds
     @Singleton
-    abstract fun bindExpensePageRepository(
-        expenseListRepository: ExpenseListFakeRepositoryImpl
-    ): ExpenseListRepository
-
-    @Binds
-    @Singleton
     abstract fun bindMemberRepository(memberRepositoryImpl: MemberRepositoryImpl): MemberRepository
-
-    @Binds
-    @Singleton
-    abstract fun bindExpenseDetailRepository(
-        expenseDetailRepositoryImpl: ExpenseDetailRepositoryImpl
-    ): ExpenseDetailRepository
-
-    @Binds
-    @Singleton
-    abstract fun bindExpenseRepository(
-        expenseRepositoryImpl: ExpenseRepositoryImpl
-    ): ExpenseRepository
-
-    @Binds
-    @Singleton
-    abstract fun bindReceiptRepository(
-        receiptRepositoryImpl: ReceiptRepositoryImpl
-    ): ReceiptRepository
 
     @Binds
     @Singleton
@@ -81,10 +47,4 @@ abstract class RepositoryModule {
     abstract fun bindAuthenticationRepository(
         authenticationRepositoryImpl: AuthenticationRepositoryImpl
     ): AuthenticationRepository
-
-    @Binds
-    @Singleton
-    abstract fun bindTransferRepository(
-        transferRepositoryImpl: TransferRepositoryImpl
-    ): TransferRepository
 }
