+++ app/src/main/java/com/kappzzang/jeongsan/di/RoomModule.kt
@@ -2,7 +2,6 @@ package com.kappzzang.jeongsan.di
 
 import android.content.Context
 import androidx.room.Room
-import com.kappzzang.jeongsan.datasource.expense.ExpenseDatabase
 import com.kappzzang.jeongsan.datasource.group.GroupDatabase
 import com.kappzzang.jeongsan.datasource.member.MemberContract
 import com.kappzzang.jeongsan.datasource.member.MemberDatabase
@@ -22,11 +21,6 @@ object RoomModule {
     @Singleton
     fun provideGroupDatabase(@ApplicationContext context: Context): GroupDatabase =
         GroupDatabase.getInstance(context)
-
-    @Provides
-    @Singleton
-    fun provideExpenseDatabase(@ApplicationContext context: Context): ExpenseDatabase =
-        ExpenseDatabase.getInstance(context)
 
     @Provides
     @Singleton
