+++ app/src/main/java/com/kappzzang/jeongsan/di/UseCaseModule.kt
@@ -1,26 +1,19 @@
 package com.kappzzang.jeongsan.di
 
-import com.kappzzang.jeongsan.repository.ExpenseDetailRepository
-import com.kappzzang.jeongsan.repository.ExpenseRepository
 import com.kappzzang.jeongsan.repository.GroupInfoRepository
 import com.kappzzang.jeongsan.repository.KakaoAuthenticationRepository
 import com.kappzzang.jeongsan.repository.MemberRepository
-import com.kappzzang.jeongsan.repository.ReceiptRepository
 import com.kappzzang.jeongsan.repository.TransferRepository
 import com.kappzzang.jeongsan.repository.UserInfoRepository
 import com.kappzzang.jeongsan.usecase.AuthenticateWithKakaoUseCase
 import com.kappzzang.jeongsan.usecase.AuthorizeWithKakaoUseCase
-import com.kappzzang.jeongsan.usecase.EditExpenseDetailUseCase
 import com.kappzzang.jeongsan.usecase.GetDoneGroupUseCase
-import com.kappzzang.jeongsan.usecase.GetExpenseDetailUseCase
-import com.kappzzang.jeongsan.usecase.GetExpenseUseCase
 import com.kappzzang.jeongsan.usecase.GetInviteInfoUseCase
 import com.kappzzang.jeongsan.usecase.GetProgressingGroupUseCase
 import com.kappzzang.jeongsan.usecase.GetTransferInfoUseCase
 import com.kappzzang.jeongsan.usecase.GetUserInfoUseCase
 import com.kappzzang.jeongsan.usecase.RegisterWithKakaoUseCase
 import com.kappzzang.jeongsan.usecase.SendTransferMessageUseCase
-import com.kappzzang.jeongsan.usecase.UploadExpenseUseCase
 import com.kappzzang.jeongsan.usecase.UploadGroupInfoUseCase
 import com.kappzzang.jeongsan.util.AuthenticationRepository
 import dagger.Module
@@ -47,22 +40,6 @@ object UseCaseModule {
     @Provides
     fun provideGetInviteInfoUseCase(memberRepository: MemberRepository): GetInviteInfoUseCase =
         GetInviteInfoUseCase(memberRepository)
-
-    @Provides
-    fun provideGetExpenseDetailUseCase(expenseDetailRepository: ExpenseDetailRepository) =
-        GetExpenseDetailUseCase(expenseDetailRepository)
-
-    @Provides
-    fun provideGetExpenseUseCase(expenseRepository: ExpenseRepository) =
-        GetExpenseUseCase(expenseRepository)
-
-    @Provides
-    fun provideEditExpenseDetailUseCase(expenseDetailRepository: ExpenseDetailRepository) =
-        EditExpenseDetailUseCase(expenseDetailRepository)
-
-    @Provides
-    fun provideUploadExpenseUseCase(receiptRepository: ReceiptRepository) =
-        UploadExpenseUseCase(receiptRepository)
 
     @Provides
     fun provideAuthenticateWithKakaoUseCase(
